video_service/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
confluent-kafka==2.6.0
cryptography==43.0.3
Django==5.1.3
django-allauth==65.2.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
pillow==11.0.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
setuptools==75.3.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
celery
django-celery-beat
django-storages
boto3
channels
channels-redis
whitenoise